refactor(py/linear): extract GraphQL queries into queries module
