+++ python/mirage/core/linear/_client.py
@@ -14,6 +14,16 @@
 
 import aiohttp
 
+from mirage.core.linear.queries import (COMMENT_CREATE_MUTATION,
+                                        COMMENT_UPDATE_MUTATION,
+                                        ISSUE_COMMENTS_QUERY,
+                                        ISSUE_CREATE_MUTATION,
+                                        ISSUE_LOOKUP_QUERY, ISSUE_QUERY,
+                                        ISSUE_SEARCH_QUERY,
+                                        ISSUE_UPDATE_MUTATION,
+                                        TEAM_CYCLES_QUERY, TEAM_ISSUES_QUERY,
+                                        TEAM_LIST_QUERY, TEAM_MEMBERS_QUERY,
+                                        TEAM_PROJECTS_QUERY, USER_LOOKUP_QUERY)
 from mirage.resource.linear.config import LinearConfig
 from mirage.resource.secrets import reveal_secret
 
@@ -100,318 +110,6 @@ async def paginate_connection(
             break
         merged_vars["after"] = page_info.get("endCursor")
     return nodes
-
-
-TEAM_LIST_QUERY = """
-query Teams($first: Int!, $after: String) {
-  teams(first: $first, after: $after) {
-    nodes {
-      id
-      key
-      name
-      description
-      timezone
-      updatedAt
-      states {
-        nodes {
-          id
-          name
-          type
-        }
-      }
-    }
-    pageInfo {
-      hasNextPage
-      endCursor
-    }
-  }
-}
-"""
-
-TEAM_MEMBERS_QUERY = """
-query TeamMembers($teamId: String!, $first: Int!, $after: String) {
-  team(id: $teamId) {
-    members(first: $first, after: $after) {
-      nodes {
-        id
-        name
-        displayName
-        email
-        active
-        admin
-        url
-        updatedAt
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"""
-
-TEAM_ISSUES_QUERY = """
-query TeamIssues($teamId: String!, $first: Int!, $after: String) {
-  team(id: $teamId) {
-    issues(first: $first, after: $after) {
-      nodes {
-        id
-        identifier
-        title
-        description
-        priority
-        url
-        createdAt
-        updatedAt
-        team {
-          id
-          key
-          name
-        }
-        state {
-          id
-          name
-        }
-        project {
-          id
-          name
-        }
-        cycle {
-          id
-          name
-          number
-        }
-        assignee {
-          id
-          name
-          email
-        }
-        creator {
-          id
-          name
-          email
-        }
-        labels {
-          nodes {
-            id
-            name
-          }
-        }
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"""
-
-TEAM_PROJECTS_QUERY = """
-query TeamProjects($teamId: String!, $first: Int!, $after: String) {
-  team(id: $teamId) {
-    projects(first: $first, after: $after) {
-      nodes {
-        id
-        name
-        description
-        status {
-          type
-        }
-        url
-        updatedAt
-        lead {
-          id
-        }
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"""
-
-TEAM_CYCLES_QUERY = """
-query TeamCycles($teamId: String!, $first: Int!, $after: String) {
-  team(id: $teamId) {
-    cycles(first: $first, after: $after) {
-      nodes {
-        id
-        name
-        number
-        startsAt
-        endsAt
-        updatedAt
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"""
-
-ISSUE_QUERY = """
-query Issue($issueId: String!) {
-  issue(id: $issueId) {
-    id
-    identifier
-    title
-    description
-    priority
-    url
-    createdAt
-    updatedAt
-    team {
-      id
-      key
-      name
-    }
-    state {
-      id
-      name
-    }
-    project {
-      id
-      name
-    }
-    cycle {
-      id
-      name
-      number
-    }
-    assignee {
-      id
-      name
-      email
-    }
-    creator {
-      id
-      name
-      email
-    }
-    labels {
-      nodes {
-        id
-        name
-      }
-    }
-  }
-}
-"""
-
-ISSUE_COMMENTS_QUERY = """
-query IssueComments($issueId: String!, $first: Int!, $after: String) {
-  issue(id: $issueId) {
-    comments(first: $first, after: $after) {
-      nodes {
-        id
-        body
-        url
-        createdAt
-        updatedAt
-        user {
-          id
-          name
-          displayName
-          email
-        }
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"""
-
-ISSUE_LOOKUP_QUERY = """
-query IssueLookup($teamKey: String!, $number: Float!) {
-  issues(
-    filter: {
-      team: { key: { eq: $teamKey } }
-      number: { eq: $number }
-    }
-    first: 1
-  ) {
-    nodes {
-      id
-      identifier
-    }
-  }
-}
-"""
-
-USER_LOOKUP_QUERY = """
-query UserLookup($email: String!) {
-  users(filter: { email: { eq: $email } }, first: 1) {
-    nodes {
-      id
-      email
-      name
-    }
-  }
-}
-"""
-
-ISSUE_CREATE_MUTATION = """
-mutation IssueCreate($input: IssueCreateInput!) {
-  issueCreate(input: $input) {
-    success
-    issue {
-      id
-      identifier
-    }
-  }
-}
-"""
-
-ISSUE_UPDATE_MUTATION = """
-mutation IssueUpdate($id: String!, $input: IssueUpdateInput!) {
-  issueUpdate(id: $id, input: $input) {
-    success
-    issue {
-      id
-      identifier
-    }
-  }
-}
-"""
-
-COMMENT_CREATE_MUTATION = """
-mutation CommentCreate($input: CommentCreateInput!) {
-  commentCreate(input: $input) {
-    success
-    comment {
-      id
-      issue {
-        id
-        identifier
-      }
-    }
-  }
-}
-"""
-
-COMMENT_UPDATE_MUTATION = """
-mutation CommentUpdate($id: String!, $input: CommentUpdateInput!) {
-  commentUpdate(id: $id, input: $input) {
-    success
-    comment {
-      id
-      issue {
-        id
-        identifier
-      }
-    }
-  }
-}
-"""
 
 
 async def list_teams(config: LinearConfig) -> list[dict]:
@@ -617,22 +315,6 @@ async def comment_update(
             if item.get("id") == comment_id:
                 return item
     return comment
-
-
-ISSUE_SEARCH_QUERY = """
-query IssueSearch($term: String!, $first: Int) {
-  searchIssues(term: $term, first: $first) {
-    nodes {
-      id
-      identifier
-      title
-      state { id name }
-      assignee { id displayName email }
-      url
-    }
-  }
-}
-"""
 
 
 async def search_issues(
